# openapi.yaml
openapi: 3.0.1
info:
  title: Mock Salesforce Object API Gateway
  version: 1.0.1
  description: Node.js gateway that routes object-based requests to Salesforce Apex REST methods using the `MultiObjectAPI` class.

servers:
  - url: https://aiactionsnodejs.onrender.com

paths:
  /get-records:
    post:
      summary: Get records by text filter from Salesforce
      operationId: fetchRecordsFromSalesforce
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tablename
                - filter
              properties:
                tablename:
                  type: string
                  example: "contact"
                filter:
                  type: string
                  example: "john"
      responses:
        "200":
          description: List of matching records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordsResponse"
        "400":
          description: Missing or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /add-record:
    post:
      summary: Add a new record to Salesforce
      operationId: createSalesforceRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/AddRecordRequest"
                - type: object
                  properties:
                    action:
                      type: string
                      example: "add"
      responses:
        "201":
          description: Record added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddResponse"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /update-record:
    post:
      summary: Update a record in Salesforce
      operationId: updateSalesforceRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateRecordRequest"
                - type: object
                  properties:
                    action:
                      type: string
                      example: "update"
      responses:
        "200":
          description: Record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /delete-record:
    post:
      summary: Delete a record in Salesforce
      operationId: deleteSalesforceRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tablename
                - Id
              properties:
                tablename:
                  type: string
                  example: "contact"
                Id:
                  type: string
                  example: "002"
                action:
                  type: string
                  example: "delete"
      responses:
        "200":
          description: Record deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /send-email:
    post:
      summary: Send an email using Salesforce Apex
      operationId: sendSalesforceEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - to
                - subject
                - text
              properties:
                to:
                  type: string
                  format: email
                  example: "recipient@example.com"
                subject:
                  type: string
                  example: "Test Email"
                text:
                  type: string
                  example: "This is a test email from the mock API server."
                action:
                  type: string
                  example: "email"
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  messageId:
                    type: string
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    Record:
      type: object
      properties:
        Id:
          type: string
        FirstName:
          type: string
        LastName:
          type: string
        Email:
          type: string

    RecordsResponse:
      type: object
      properties:
        count:
          type: integer
        records:
          type: array
          items:
            $ref: "#/components/schemas/Record"
        message:
          type: string

    AddRecordRequest:
      type: object
      required:
        - tablename
        - FirstName
        - LastName
        - Email
      properties:
        tablename:
          type: string
          example: "lead"
        FirstName:
          type: string
        LastName:
          type: string
        Email:
          type: string

    AddResponse:
      type: object
      properties:
        message:
          type: string
        record:
          $ref: "#/components/schemas/Record"

    UpdateRecordRequest:
      type: object
      required:
        - tablename
        - Id
      properties:
        tablename:
          type: string
        Id:
          type: string
        FirstName:
          type: string
        LastName:
          type: string
        Email:
          type: string

    UpdateResponse:
      type: object
      properties:
        message:
          type: string
        record:
          $ref: "#/components/schemas/Record"

    DeleteResponse:
      type: object
      properties:
        message:
          type: string
        deleted:
          $ref: "#/components/schemas/Record"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string